requirements.txt
# Perun's BlackBook - FastAPI Application Dependencies
# Install with: pip install -r requirements.txt
# Note: Using flexible versions for Python 3.14 compatibility

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Environment & Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Security & Encryption
cryptography>=42.0.0

# Google OAuth & API
google-auth>=2.27.0,<2.42.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.114.0

# Templates
jinja2>=3.1.3

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0
requests>=2.31.0

# AI Providers (Phase 5)
anthropic>=0.40.0
google-genai>=1.0.0

# Background Tasks (Phase 6)
apscheduler>=3.10.0
